use only smart_open
