+++ plynx/__init__.py
@@ -4,4 +4,4 @@ PLynx
 Interactive, Scalable, Shareable and Reproducible Workflow Orchestration framework
 """
 
-__version__ = '1.7.4'
+__version__ = '1.7.5'
