+++ plynx/utils/file_handler.py
@@ -1,14 +1,25 @@
-import io
+import smart_open
 import uuid
-from plynx.utils.remote import get_driver
+
+from plynx.utils.config import get_storage_config
+
+_global_storage_config = None
+
+
+def get_global_storage_config():
+    global _global_storage_config
+    if not _global_storage_config:
+        _global_storage_config = get_storage_config()
+    return _global_storage_config
+
+
+def open(file, mode="rt"):
+    file = f"{get_global_storage_config().prefix}{file}"
+    return smart_open.open(file, mode)
 
 
 def get_file_stream(file_path, preview=False, file_type=None):
-    content = get_driver().get_contents_handler(file_path)
-    content_stream = io.BytesIO()
-    content.get_contents_to_file(content_stream)
-    content_stream.seek(0)
-    return content_stream
+    return open(file_path, "rb")
 
 
 def upload_file_stream(fp, file_path=None, seek=True):
@@ -16,11 +27,6 @@ def upload_file_stream(fp, file_path=None, seek=True):
         fp.seek(0)
     if file_path is None:
         file_path = str(uuid.uuid1())
-    content = get_driver().get_contents_handler(file_path)
-    content.set_contents_from_file(fp)
+    with open(file_path, "wb") as fo:
+        fo.write(fp)
     return file_path
-
-
-def remove(file_path):
-    content = get_driver().get_contents_handler(file_path)
-    content.remove()
